Not Currently Working
RandomMathBehavior and SubtractOriginalValue not working with new version of delegateMath that only has int x.  No longer can modify the buffer.  Need to figure out how to store and read the original value and pass it through the pipeline.  Look at swen300-labs repo and lab02 branch has working version that was submitted to gradescope.  Trying to fix issues that I lost points on.